work on PlayerHand Class
Remove function: removes cards from player's hand and throws then in order if the thrown cards are a sequence. Updates point, and restarts track of a player's marked cards.

evaluateLegalCombination function: recives an array of cards to remove, returns true if the array is a legal combination of card to throw.

add function: extends super.add. updates a player points in hand

updateMarkedCards(card): recives a card in the players hand. If the card is marked /unmarked it is added / removed from the marked cards array property of the players hand